% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{antarctic_bathy}
\alias{antarctic_bathy}
\title{Antarctic Stereographic bathymetry}
\format{
SpatialPolygonsDataFrame bathymetry in Antarctic Stereographic projection (+init=epsg:3031). Contains latitudes between -80 and -30 degrees. Generalized from ETOPO1 grid.
}
\source{
\href{https://doi.org/10.7289/V5C8276M}{ETOPO1 1 Arc-Minute Global Relief Model}
}
\usage{
antarctic_bathy
}
\description{
Antarctic Stereographic bathymetry
}
\seealso{
\code{\link{basemap}}

Other shapefiles: 
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}},
\code{\link{fishingAreasNor}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
