% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic_bathy}
\alias{arctic_bathy}
\title{Arctic Stereographic bathymetry}
\format{
SpatialPolygonsDataFrame bathymetry in Arctic Stereographic projection (+init=epsg:3995). Contains latitudes between 30 and 90 degrees. eneralized from ETOPO1 grid.
}
\source{
\href{https://doi.org/10.7289/V5C8276M}{ETOPO1 1 Arc-Minute Global Relief Model}
}
\usage{
arctic_bathy
}
\description{
Arctic Stereographic bathymetry
}
\seealso{
\code{\link{basemap}}

Other shapefiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}},
\code{\link{fishingAreasNor}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
