% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_limits.R
\name{auto_limits}
\alias{auto_limits}
\title{Automatic limits for basemap}
\usage{
auto_limits(
  data,
  lon = NULL,
  lat = NULL,
  proj.in = "+init=epsg:4326",
  proj.out = NULL,
  expand.factor = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame containing data for which the limits should be calculated.}

\item{lon, lat}{Names of longitude and latitude columns in \code{data} as character or integer index. If \code{NULL}, the column names are \link[=guess_coordinate_columns]{guessed}.}

\item{proj.in}{Original \code{\link[sp]{proj4string}} projection. Must be defined as character argument.}

\item{proj.out}{Resulting map projection. See \code{\link{transform_coord}}.}

\item{expand.factor}{Expansion factor for map limits. Set to \code{NULL} to ignore.}

\item{verbose}{if \code{TRUE}, the function prints information about the changed projection. Switch to \code{FALSE} to make the function silent.}

\item{rotate}{Logical indicating whether the limits should be rotated to point towards the pole relative to mid-longitude limit.}
}
\description{
An internal function to make \code{\link{basemap}} code more readable
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function. The function does some funky illogical action (seeing the argument names) when feeding in projected coordinates. This is to make the automatic limits to work for projected coordinates too.
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
