% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_data.R
\name{basemap_data}
\alias{basemap_data}
\title{Create basemapData object for basemap plotting}
\usage{
basemap_data(
  limits = NULL,
  data = NULL,
  shapefiles = NULL,
  bathymetry = FALSE,
  glaciers = FALSE,
  resolution = "low",
  lon.interval = NULL,
  lat.interval = NULL,
  expand.factor = 1.1,
  rotate = FALSE
)
}
\arguments{
\item{limits}{See \code{\link{basemap}}}

\item{data}{See \code{\link{basemap}}}

\item{shapefiles}{See \code{\link{basemap}}}

\item{bathymetry}{See \code{\link{basemap}}}

\item{glaciers}{See \code{\link{basemap}}}

\item{resolution}{See \code{\link{basemap}}}

\item{lon.interval, lat.interval}{See \code{\link{basemap}}}

\item{expand.factor}{See \code{\link{auto_limits}}}
}
\description{
Internal function to create a \code{basemapData} object for \code{\link{basemap}}
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function. Common users do not need to worry about these details.
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
