% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{bs_glacier}
\alias{bs_glacier}
\title{Barents Sea glacier shapes.}
\format{
SpatialPolygonsDataFrame glacier shapes for the Barents Sea in UTM 36N projection (+init=epsg:32636). Obtained from Natural Earth Data (10m vectors).
}
\source{
\href{http://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
bs_glacier
}
\description{
Barents Sea glacier shapes.
}
\seealso{
\code{\link{basemap}}

Other shapefiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}},
\code{\link{fishingAreasNor}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
