% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{bs_land}
\alias{bs_land}
\title{Barents Sea land shapes.}
\format{
SpatialPolygonsDataFrame land shapes for the Barents Sea in UTM 36N projection (+init=epsg:32636). Obtained from Natural Earth Data (10m vectors). Includes the islands dataset.
}
\source{
\href{http://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
bs_land
}
\description{
Barents Sea land shapes.
}
\seealso{
\code{\link{basemap}}

Other shapefiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}},
\code{\link{fishingAreasNor}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
