% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{dd_clip_boundary}
\alias{dd_clip_boundary}
\title{Create clip boundary from decimal degree limits}
\usage{
dd_clip_boundary(limits, crs, expand.factor = NULL)
}
\arguments{
\item{limits}{Numeric vector of length 4 giving the map limits in decimal degrees. See \link{basemap}.}

\item{crs}{Coordinate reference system in \code{\link[sf]{st_crs}} format.}

\item{expand.factor}{Expansion factor for map limits. Set to \code{NULL} to ignore.}
}
\description{
Creates a clip boundary for map cropping from decimal degree limits 
counter-clockwise following parallels
}
\author{
Mikko Vihtakari
}
\keyword{internal}
