% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{dd_rbathy}
\alias{dd_rbathy}
\title{Decimal degree bathymetry}
\format{
\link[stars:read_stars]{Raster} bathymetry in decimal degrees (EPSG:4326). Downsampled from ETOPO 60 arc-second grid.
}
\source{
NOAA National Centers for Environmental Information. 2022: ETOPO 2022 15 Arc-Second Global Relief Model. NOAA National Centers for Environmental Information. \doi{10.25921/fd45-gt74}
}
\usage{
dd_rbathy
}
\description{
Decimal degree bathymetry
}
\seealso{
Other mapfiles: 
\code{\link{dd_land}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
