% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{dd_to_deg}
\alias{dd_to_deg}
\title{Convert decimal degree values to angular degrees}
\usage{
dd_to_deg(x)
}
\arguments{
\item{x}{numeric to be converted}
}
\value{
A vector of angular degrees
}
\description{
Converts decimal degree values to angular degrees. Used in decimal degree limit calculations.
}
\seealso{
Other degree converters: 
\code{\link{deg_to_dd}()}
}
\author{
Mikko Vihtakari
}
\concept{degree converters}
\keyword{internal}
