% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{define_bathy_style}
\alias{define_bathy_style}
\title{Define bathymetry style for \code{\link{basemap}}}
\usage{
define_bathy_style(x)
}
\arguments{
\item{x}{Character argument giving the input bathymetry style. Partially matched and can be abbreviated. See \code{bathy.style} in \code{\link{basemap}}.}
}
\value{
Returns a named character vector with the name pointing to the bathymetry style and value to internal map element command for \code{\link{basemap}} (see \code{\link{map_cmd}}).
}
\description{
Defines bathymetry style to be used in \code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
