% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{deg_to_dd}
\alias{deg_to_dd}
\title{Convert angular degrees to decimal degree values}
\usage{
deg_to_dd(x)
}
\arguments{
\item{x}{numeric to be converted}
}
\value{
A vector of decimal degrees
}
\description{
Converts angular degree values to decimal degrees. Used in decimal degree limit calculations.
}
\seealso{
Other degree converters: 
\code{\link{dd_to_deg}()}
}
\author{
Mikko Vihtakari
}
\concept{degree converters}
\keyword{internal}
