% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{fishingAreasNor}
\alias{fishingAreasNor}
\title{Major fishing zones of Norway}
\format{
SpatialPolygonsDataFrame in decimal degrees (+init=epsg:4326) containing major fishing zones defined by the Norwegian Directorate of Fisheries.
}
\source{
\href{https://kart.fiskeridir.no/stat}{Norwegian Directorate of Fisheries}
}
\usage{
fishingAreasNor
}
\description{
Major fishing zones of Norway
}
\seealso{
\code{\link{basemap}}

Other shapefiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
