% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{FS}
\alias{FS}
\title{Convert font sizes measured as points to ggplot font sizes}
\usage{
FS(x)
}
\arguments{
\item{x}{numeric vector giving the font sizes in points}
}
\value{
Returns a numeric vector of lenght \code{x} of ggplot font sizes
}
\description{
Converts font sizes measured as points (as given by most programs such as MS Word etc.) to ggplot font sizes
}
\seealso{
Other size adjustors: 
\code{\link{LS}()}
}
\concept{size adjustors}
\keyword{internal}
