% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{getCores}
\alias{getCores}
\title{Pick a suitable number of cores}
\usage{
getCores()
}
\value{
Integer of suitable number of cores
}
\description{
Picks maximum four cores for parallel processing
}
\author{
The \href{https://github.com/StoXProject/RstoxData/blob/master/R/Utilities.R}{StoXProject}
}
\keyword{internal}
