% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_coordinate_columns.R
\name{guess_coordinate_columns}
\alias{guess_coordinate_columns}
\title{Guess longitude and latitude columns in a data frame}
\usage{
guess_coordinate_columns(data, lon = NULL, lat = NULL)
}
\arguments{
\item{data}{Dataframe containing data for which the limits should be calculated.}

\item{lon, lat}{Character defining the name of the longitude and latitude columns in \code{data}. Use \code{NULL} to guess the longitude and/or latitude columns in \code{x}.}
}
\value{
A named vector of colummn names.
}
\description{
An internal function to make \code{\link{basemap}} code more readable
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function.
}
\seealso{
\code{\link{auto_limits}}, \code{\link{transform_coord}}, \code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
