% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\encoding{UTF-8}
\name{ices_areas}
\alias{ices_areas}
\title{ICES Advisory Areas}
\format{
\code{\link[sf:st_sf]{sf object}} containing ICES Advisory Areas.
}
\source{
\href{https://www.ices.dk/}{International Council for the Exploration of the Sea}
}
\usage{
ices_areas
}
\description{
ICES Advisory Areas
}
\examples{
if(requireNamespace("ggspatial")) {
\donttest{
basemap(ices_areas) + 
ggspatial::annotation_spatial(ices_areas, fill = NA)
}
}
}
\seealso{
Other datasets: 
\code{\link{fdir_main_areas}},
\code{\link{fdir_sub_areas}}
}
\concept{datasets}
\keyword{datasets}
\keyword{fishery}
\keyword{shapefiles}
