% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_decimal_limit.R
\name{is_decimal_limit}
\alias{is_decimal_limit}
\title{Test whether a limit argument is specified as decimal degrees.}
\usage{
is_decimal_limit(limits)
}
\arguments{
\item{limits}{A numeric vector of length 1 or 4. See \code{\link{basemap}}}
}
\value{
A logical value
}
\description{
An internal function to make \code{\link{basemap}} code more readable
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function.
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
