% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_bathymetry.R
\name{ksmooth_polys}
\alias{ksmooth_polys}
\title{Wrapper to \code{\link[smoothr]{smooth_ksmooth}} SpatialPolygonsDataFrames}
\usage{
ksmooth_polys(x, k, N)
}
\arguments{
\item{x}{SpatialPolygonsDataFrame}

\item{k}{The \code{smoothness} parameter in \code{\link[smoothr]{smooth_ksmooth}}}

\item{N}{The \code{n} parameter in \code{\link[smoothr]{smooth_ksmooth}}}
}
\value{
Smoothed \code{SpatialPolygonsDataFrame}
}
\description{
Wrapper to \code{\link[smoothr]{smooth_ksmooth}} SpatialPolygonsDataFrames
}
\keyword{internal}
