% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{LS}
\alias{LS}
\title{Convert line sizes measured as points to ggplot line sizes}
\usage{
LS(x)
}
\arguments{
\item{x}{numeric vector giving the lines sizes in points}
}
\value{
Returns a numeric vector of lenght \code{x} of ggplot line sizes
}
\description{
Converts line sizes measured as points (as given by most programs such as Adobe Illustrator etc.) to ggplot font sizes
}
\keyword{internal}
