% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{qmap}
\alias{qmap}
\title{Quick map}
\usage{
qmap(
  data,
  x = NULL,
  y = NULL,
  geom = "point",
  limits = NULL,
  bathymetry = FALSE,
  bathy.style = "poly_blues",
  ...
)
}
\arguments{
\item{data}{Data frame to use.}

\item{x, y, ...}{Aesthetics passed into each layer. Longitude and latitude columns are automatically recognized using the \code{\link{guess_coordinate_columns}} function.}

\item{geom}{Character vector specifying geom(s) to draw. Defaults to "point". Other \code{geom}s have not been implemented yet.}

\item{limits}{Map limits. See the \code{limits} argument in \code{\link{basemap}}. If \code{NULL} the limits are automatically taken from \code{data}}

\item{bathymetry}{Logical indicating whether bathymetry should be added to the map.}

\item{bathy.style}{Character defining the style for bathymetry contours. See the \code{bathy.style} argument in \code{\link{basemap}}.}
}
\description{
\code{qmap} is a shortcut similar to ggplot2's \code{\link[ggplot2]{qplot}} designed to quickly plot data with a limited range of options.
}
\author{
Mikko Vihtakari
}
