% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rotate_crs}
\alias{rotate_crs}
\title{Rotate CRS}
\usage{
rotate_crs(crs, meridians)
}
\arguments{
\item{crs}{The CRS to be rotated in \code{\link[sf]{st_crs}} format}

\item{meridians}{A numeric vector giving the two meridians which should be used in the rotation}
}
\value{
Rotated CRS in \code{\link[sf]{st_crs}} format
}
\description{
Rotates a CRS such that North is in the middle of two meridians
}
\author{
Mikko Vihtakari
}
\keyword{internal}
