% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or \code{\link{round}}}
}
\value{
Rounded numeric vector
}
\description{
Round to multiple of any number
}
\author{
Hadley Wickham
}
\keyword{internal}
