% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coord.R
\name{transform_coord}
\alias{transform_coord}
\title{Transform spatial coordinates to another projection}
\usage{
transform_coord(
  x = NULL,
  lon = NULL,
  lat = NULL,
  new.names = "auto",
  proj.in = "+init=epsg:4326",
  proj.out = NULL,
  verbose = FALSE,
  bind = FALSE,
  na = "ignore"
)
}
\arguments{
\item{x}{Data frame to be transformed. Can be omitted if numeric vectors are assigned to \code{lon} and \code{lat}.}

\item{lon, lat}{Either a name of the longitude and latitude columns in \code{x} or a numeric vector containing longitude and latitude coordinates. Use \code{NULL} to \link[=guess_coordinate_columns]{guess the longitude and/or latitude columns} in \code{x}.}

\item{new.names}{Character vector of length 2 specifying the names of transformed longitude and latitude columns, respectively. Alternatively \code{NULL}, which returns column names from \code{x} or "auto", which uses \code{NULL} if \code{bind = FALSE} and \code{c("lon.proj", "lat.proj")} if \code{bind = TRUE}.}

\item{proj.in}{Original \code{\link[sp:is.projected]{proj4string}} projection. If \code{NULL}, the projection is taken from \code{x}. \code{x} must be a \link[sp:Spatial-class]{Spatial} object in that case.}

\item{proj.out}{Character. Either \code{NULL}, \code{\link[sp:is.projected]{proj4string}} projection the coordinates should be transformed to or a name of shapefiles in \code{\link{shapefile_list}}. If \code{NULL}, the output projection will be automatically determined from data. This option requires decimal degrees as input option.}

\item{verbose}{Logical indicating whether information about the projection should be returned as message. Set to \code{FALSE} to make the function silent.}

\item{bind}{logical. Should only transformed coordinates be returned (\code{FALSE}, default) or should x be returned with transformed coordinates (\code{TRUE})?}

\item{na}{character specifying the NA action for missing coordinates. The "ignore" option ignores the coordinates and returns NAs to transformed coordinates. The "remove" option removes missing values from \code{x} returning a message while doing it. Any other character argument will trigger \code{na.fail} stopping the function in case of missing coordinates.}
}
\value{
Returns a data frame with transformed spatial coordinates.
}
\description{
Transforms spatial coordinates from original projection (decimal degrees assumed) to another projection.
}
\details{
If \code{x} is specified, the function guesses longitude and latitude columns from \code{x} by default.
}
\examples{
# Coordinates are automatically transformed to the pre-made shapefile 
# projections:
dt <- data.frame(lon = c(-150, 150), lat = c(60, 90))
transform_coord(dt)
transform_coord(dt, bind = TRUE)

dt <- data.frame(lon = c(-150, 150), lat = c(20, 50))
transform_coord(dt, bind = TRUE) # no transformation required.
}
\seealso{
Other basemap functions: 
\code{\link{basemap}()},
\code{\link{qmap}()},
\code{\link{shapefile_list}()}
}
\author{
Mikko Vihtakari
}
\concept{basemap functions}
