% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\encoding{UTF-8}
\name{dd_land}
\alias{dd_land}
\title{Decimal degree land shapes}
\format{
\link[sf:st_sf]{Simple feature collection} land shapes in decimal degrees (EPSG:4326). Obtained from Natural Earth Data (10m vectors). Includes the islands dataset.
}
\source{
\href{https://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
dd_land
}
\description{
Decimal degree land shapes
}
\seealso{
Other mapfiles: 
\code{\link{dd_rbathy}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
