% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_shapefiles.R
\name{define_shapefiles}
\alias{define_shapefiles}
\title{Define a shapefile to use in plotting from the limits argument}
\usage{
define_shapefiles(limits, force_dd = FALSE)
}
\arguments{
\item{limits}{A numeric vector of length 4: The first element defines the minimum longitude, the second element the maximum longitude, the third element the minimum latitude and the fourth element the maximum latitude of the bounding box.}

\item{force_dd}{Logical indicating whether the shapefile should be forced to DecimalDegree. Required for transforming dd_* shapefile objects to other projections.}
}
\value{
A list containing the correct shapefile, a logical statement whether the limits were supplied as decimal degrees and coordinate reference system.
}
\description{
An internal function to make \code{\link{basemap}} code more readable
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function.
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
