% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{quiet}
\alias{quiet}
\title{Return function output quietly}
\usage{
quiet(x)
}
\arguments{
\item{x}{function}
}
\value{
Output of \code{x}
}
\description{
Returns function output without printed \code{\link{cat}} messages
}
\author{
Hadley Wickham
}
\keyword{internal}
