% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{select_element}
\alias{select_element}
\title{Select an element of each vector from a list}
\usage{
select_element(x, y)
}
\arguments{
\item{x}{list}

\item{y}{number of element. Must be integer}
}
\value{
The selected element from the list
}
\description{
Selects y'th element of each vector from a list
}
\keyword{internal}
