% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{A ggplot2 theme for maps}
\usage{
theme_map(..., grid.col, grid.size)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[ggplot2]{ggtheme}}.}

\item{grid.col}{Character code specifying the color of grid lines. Use \code{NA} to remove the grid lines.}

\item{grid.size}{Numeric value specifying the width of grid lines.}
}
\value{
A ggplot2 theme layer.
}
\description{
A ggplot2 theme for maps.
}
\seealso{
Other customize shapefiles: 
\code{\link{auto_limits}()},
\code{\link{reorder_layers}()}
}
\concept{customize shapefiles}
